requirements.txt
# 기본 웹 서버 프레임워크
fastapi>=0.121.0
uvicorn>=0.38.0

# 환경변수 관리
python-dotenv>=1.2.1

# AI 에이전트관련
langchain>=1.0.5
openai>=2.7.1
pydantic>=2.12.4

# 데이터 처리

# HTTP 요청용
requests>=2.32.5

# 기타 (로그, 로컬DB등)

annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0

h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0

langchain-core>=1.0.4
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41

orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0

pydantic_core>=2.41.5

PyYAML==6.0.3

requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.49.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0

xxhash==3.6.0
zstandard==0.25.0